GatewayMetaState.applyClusterStateUpdaters
