+++ server/src/main/java/org/elasticsearch/gateway/Gateway.java
@@ -31,8 +31,10 @@ import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.discovery.zen.ElectMasterService;
 import org.elasticsearch.index.Index;
+import org.elasticsearch.indices.IndicesService;
 
 import java.util.Arrays;
+import java.util.function.Function;
 
 public class Gateway {
 
@@ -43,9 +45,12 @@ public class Gateway {
     private final TransportNodesListGatewayMetaState listGatewayMetaState;
 
     private final int minimumMasterNodes;
+    private final IndicesService indicesService;
 
     public Gateway(final Settings settings, final ClusterService clusterService,
-                   final TransportNodesListGatewayMetaState listGatewayMetaState) {
+                   final TransportNodesListGatewayMetaState listGatewayMetaState,
+                   final IndicesService indicesService) {
+        this.indicesService = indicesService;
         this.clusterService = clusterService;
         this.listGatewayMetaState = listGatewayMetaState;
         this.minimumMasterNodes = ElectMasterService.DISCOVERY_ZEN_MINIMUM_MASTER_NODES_SETTING.get(settings);
@@ -119,7 +124,12 @@ public class Gateway {
                 }
             }
         }
-        listener.onSuccess(ClusterState.builder(clusterService.getClusterName()).metaData(metaDataBuilder).build());
+        ClusterState recoveredState = Function.<ClusterState>identity()
+            .andThen(state -> ClusterStateUpdaters.upgradeAndArchiveUnknownOrInvalidSettings(state, clusterService.getClusterSettings()))
+            .andThen(state -> ClusterStateUpdaters.closeBadIndices(state, indicesService))
+            .apply(ClusterState.builder(clusterService.getClusterName()).metaData(metaDataBuilder).build());
+
+        listener.onSuccess(recoveredState);
     }
 
     public interface GatewayStateRecoveredListener {
