+++ server/src/main/java/org/elasticsearch/gateway/GatewayMetaState.java
@@ -34,16 +34,18 @@ import org.elasticsearch.cluster.metadata.Manifest;
 import org.elasticsearch.cluster.metadata.MetaData;
 import org.elasticsearch.cluster.metadata.MetaDataIndexUpgradeService;
 import org.elasticsearch.cluster.node.DiscoveryNode;
-import org.elasticsearch.cluster.node.DiscoveryNodes;
 import org.elasticsearch.cluster.routing.RoutingNode;
 import org.elasticsearch.cluster.routing.ShardRouting;
+import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.collect.ImmutableOpenMap;
 import org.elasticsearch.common.collect.Tuple;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.env.NodeEnvironment;
 import org.elasticsearch.index.Index;
+import org.elasticsearch.indices.IndicesService;
 import org.elasticsearch.plugins.MetaDataUpgrader;
+import org.elasticsearch.transport.TransportService;
 
 import java.io.IOException;
 import java.nio.file.DirectoryStream;
@@ -58,6 +60,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.function.UnaryOperator;
 
 import static org.elasticsearch.gateway.GatewayService.STATE_NOT_RECOVERED_BLOCK;
@@ -76,25 +79,35 @@ public class GatewayMetaState implements ClusterStateApplier, CoordinationState.
     protected static final Logger logger = LogManager.getLogger(GatewayMetaState.class);
 
     private final NodeEnvironment nodeEnv;
-    protected final MetaStateService metaStateService;
+    private final MetaStateService metaStateService;
     private final Settings settings;
+    private final ClusterService clusterService;
+    private final IndicesService indicesService;
+    private final TransportService transportService;
 
     //there is a single thread executing updateClusterState calls, hence no volatile modifier
     protected Manifest previousManifest;
     protected ClusterState previousClusterState;
+    protected boolean clusterStateUpdatersApplied;
     protected boolean incrementalWrite;
 
     public GatewayMetaState(Settings settings, NodeEnvironment nodeEnv, MetaStateService metaStateService,
-                            MetaDataIndexUpgradeService metaDataIndexUpgradeService, MetaDataUpgrader metaDataUpgrader) throws IOException {
+                            MetaDataIndexUpgradeService metaDataIndexUpgradeService, MetaDataUpgrader metaDataUpgrader,
+                            TransportService transportService, ClusterService clusterService,
+                            IndicesService indicesService) throws IOException {
         this.settings = settings;
         this.nodeEnv = nodeEnv;
         this.metaStateService = metaStateService;
+        this.transportService = transportService;
+        this.clusterService = clusterService;
+        this.indicesService = indicesService;
 
         ensureNoPre019State(); //TODO remove this check, it's Elasticsearch version 7 already
         ensureAtomicMoveSupported(); //TODO move this check to NodeEnvironment, because it's related to all types of metadata
         upgradeMetaData(metaDataIndexUpgradeService, metaDataUpgrader);
         initializeClusterState(ClusterName.CLUSTER_NAME_SETTING.get(settings));
         incrementalWrite = false;
+        clusterStateUpdatersApplied = false;
     }
 
     private void initializeClusterState(ClusterName clusterName) throws IOException {
@@ -113,11 +126,17 @@ public class GatewayMetaState implements ClusterStateApplier, CoordinationState.
         logger.debug("took {} to load state", TimeValue.timeValueMillis(TimeValue.nsecToMSec(System.nanoTime() - startNS)));
     }
 
-    public void setLocalNode(DiscoveryNode localNode) {
-        assert previousClusterState.nodes().getLocalNode() == null : "setLocalNode must only be called once";
-        previousClusterState = ClusterState.builder(previousClusterState)
-                .nodes(DiscoveryNodes.builder().add(localNode).localNodeId(localNode.getId()).build())
-                .build();
+    public void applyClusterStateUpdaters() {
+        assert clusterStateUpdatersApplied == false : "applyClusterStateUpdaters must only be called once";
+
+        previousClusterState = Function.<ClusterState>identity()
+            .andThen(state -> ClusterStateUpdaters.setLocalNode(state, transportService.getLocalNode()))
+            .andThen(state -> ClusterStateUpdaters.upgradeAndArchiveUnknownOrInvalidSettings(state, clusterService.getClusterSettings()))
+            .andThen(state -> ClusterStateUpdaters.closeBadIndices(state, indicesService))
+            .andThen(ClusterStateUpdaters::recoverClusterBlocks)
+            .apply(previousClusterState);
+
+        clusterStateUpdatersApplied = true;
     }
 
     protected void upgradeMetaData(MetaDataIndexUpgradeService metaDataIndexUpgradeService, MetaDataUpgrader metaDataUpgrader)
@@ -203,7 +222,7 @@ public class GatewayMetaState implements ClusterStateApplier, CoordinationState.
 
     @Override
     public ClusterState getLastAcceptedState() {
-        assert previousClusterState.nodes().getLocalNode() != null : "Call setLocalNode before calling this method";
+        assert clusterStateUpdatersApplied : "Call applyClusterStateUpdaters before calling this method";
         return previousClusterState;
     }
