+++ server/src/main/java/org/elasticsearch/gateway/GatewayService.java
@@ -19,7 +19,6 @@
 
 package org.elasticsearch.gateway;
 
-import com.carrotsearch.hppc.cursors.ObjectCursor;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.message.ParameterizedMessage;
@@ -29,29 +28,22 @@ import org.elasticsearch.cluster.ClusterStateListener;
 import org.elasticsearch.cluster.ClusterStateUpdateTask;
 import org.elasticsearch.cluster.block.ClusterBlock;
 import org.elasticsearch.cluster.block.ClusterBlockLevel;
-import org.elasticsearch.cluster.block.ClusterBlocks;
 import org.elasticsearch.cluster.coordination.Coordinator;
-import org.elasticsearch.cluster.metadata.IndexMetaData;
-import org.elasticsearch.cluster.metadata.MetaData;
 import org.elasticsearch.cluster.node.DiscoveryNodes;
-import org.elasticsearch.cluster.routing.RoutingTable;
 import org.elasticsearch.cluster.routing.allocation.AllocationService;
 import org.elasticsearch.cluster.service.ClusterService;
 import org.elasticsearch.common.component.AbstractLifecycleComponent;
 import org.elasticsearch.common.inject.Inject;
-import org.elasticsearch.common.settings.ClusterSettings;
 import org.elasticsearch.common.settings.Setting;
 import org.elasticsearch.common.settings.Setting.Property;
 import org.elasticsearch.common.settings.Settings;
 import org.elasticsearch.common.unit.TimeValue;
 import org.elasticsearch.common.util.concurrent.AbstractRunnable;
 import org.elasticsearch.discovery.Discovery;
-import org.elasticsearch.index.Index;
 import org.elasticsearch.indices.IndicesService;
 import org.elasticsearch.rest.RestStatus;
 import org.elasticsearch.threadpool.ThreadPool;
 
-import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Function;
 
@@ -83,8 +75,6 @@ public class GatewayService extends AbstractLifecycleComponent implements Cluste
     private final AllocationService allocationService;
 
     private final ClusterService clusterService;
-
-    private final IndicesService indicesService;
 
     private final TimeValue recoverAfterTime;
     private final int recoverAfterNodes;
@@ -134,11 +124,10 @@ public class GatewayService extends AbstractLifecycleComponent implements Cluste
             recoveryRunnable = () ->
                     clusterService.submitStateUpdateTask("local-gateway-elected-state", new RecoverStateUpdateTask());
         } else {
-            final Gateway gateway = new Gateway(settings, clusterService, listGatewayMetaState);
+            final Gateway gateway = new Gateway(settings, clusterService, listGatewayMetaState, indicesService);
             recoveryRunnable = () ->
                     gateway.performStateRecovery(new GatewayRecoveryListener());
         }
-        this.indicesService = indicesService;
     }
 
     @Override
@@ -250,123 +239,13 @@ public class GatewayService extends AbstractLifecycleComponent implements Cluste
         logger.info("metadata state not restored, reason: {}", message);
     }
 
-    static class Updaters {
-        static ClusterState upgradeAndArchiveUnknownOrInvalidSettings(final ClusterState clusterState,
-                                                                      final ClusterSettings clusterSettings) {
-            final MetaData.Builder metaDataBuilder = MetaData.builder(clusterState.metaData());
-
-            metaDataBuilder.persistentSettings(
-                    clusterSettings.archiveUnknownOrInvalidSettings(
-                            clusterSettings.upgradeSettings(metaDataBuilder.persistentSettings()),
-                            e -> logUnknownSetting("persistent", e),
-                            (e, ex) -> logInvalidSetting("persistent", e, ex)));
-            metaDataBuilder.transientSettings(
-                    clusterSettings.archiveUnknownOrInvalidSettings(
-                            clusterSettings.upgradeSettings(metaDataBuilder.transientSettings()),
-                            e -> logUnknownSetting("transient", e),
-                            (e, ex) -> logInvalidSetting("transient", e, ex)));
-            return ClusterState.builder(clusterState).metaData(metaDataBuilder).build();
-        }
-
-        private static void logUnknownSetting(final String settingType, final Map.Entry<String, String> e) {
-            logger.warn("ignoring unknown {} setting: [{}] with value [{}]; archiving", settingType, e.getKey(), e.getValue());
-        }
-
-        private static void logInvalidSetting(final String settingType, final Map.Entry<String, String> e,
-                                              final IllegalArgumentException ex) {
-            logger.warn(() -> new ParameterizedMessage("ignoring invalid {} setting: [{}] with value [{}]; archiving",
-                    settingType, e.getKey(), e.getValue()), ex);
-        }
-
-        static ClusterState recoverClusterBlocks(final ClusterState state) {
-            final ClusterBlocks.Builder blocks = ClusterBlocks.builder().blocks(state.blocks());
-
-            if (MetaData.SETTING_READ_ONLY_SETTING.get(state.metaData().settings())) {
-                blocks.addGlobalBlock(MetaData.CLUSTER_READ_ONLY_BLOCK);
-            }
-
-            if (MetaData.SETTING_READ_ONLY_ALLOW_DELETE_SETTING.get(state.metaData().settings())) {
-                blocks.addGlobalBlock(MetaData.CLUSTER_READ_ONLY_ALLOW_DELETE_BLOCK);
-            }
-
-            for (final IndexMetaData indexMetaData : state.metaData()) {
-                blocks.addBlocks(indexMetaData);
-            }
-
-            return ClusterState.builder(state).blocks(blocks).build();
-        }
-
-        static ClusterState closeBadIndices(final ClusterState clusterState, final IndicesService indicesService) {
-            final MetaData.Builder builder = MetaData.builder(clusterState.metaData()).removeAllIndices();
-
-            for (IndexMetaData metaData : clusterState.metaData()) {
-                try {
-                    if (metaData.getState() == IndexMetaData.State.OPEN) {
-                        // verify that we can actually create this index - if not we recover it as closed with lots of warn logs
-                        indicesService.verifyIndexMetadata(metaData, metaData);
-                    }
-                } catch (final Exception e) {
-                    final Index electedIndex = metaData.getIndex();
-                    logger.warn(() -> new ParameterizedMessage("recovering index {} failed - recovering as closed", electedIndex), e);
-                    metaData = IndexMetaData.builder(metaData).state(IndexMetaData.State.CLOSE).build();
-                }
-                builder.put(metaData, false);
-            }
-
-            return ClusterState.builder(clusterState).metaData(builder).build();
-        }
-
-        static ClusterState updateRoutingTable(final ClusterState state) {
-            // initialize all index routing tables as empty
-            final RoutingTable.Builder routingTableBuilder = RoutingTable.builder(state.routingTable());
-            for (final ObjectCursor<IndexMetaData> cursor : state.metaData().indices().values()) {
-                routingTableBuilder.addAsRecovery(cursor.value);
-            }
-            // start with 0 based versions for routing table
-            routingTableBuilder.version(0);
-            return ClusterState.builder(state).routingTable(routingTableBuilder.build()).build();
-        }
-
-        static ClusterState removeStateNotRecoveredBlock(final ClusterState state) {
-            return ClusterState.builder(state)
-                    .blocks(ClusterBlocks.builder()
-                            .blocks(state.blocks()).removeGlobalBlock(STATE_NOT_RECOVERED_BLOCK).build())
-                    .build();
-        }
-
-        static ClusterState mixCurrentStateAndRecoveredState(final ClusterState currentState, final ClusterState recoveredState) {
-            assert currentState.metaData().indices().isEmpty();
-
-            final ClusterBlocks.Builder blocks = ClusterBlocks.builder()
-                    .blocks(currentState.blocks())
-                    .blocks(recoveredState.blocks());
-
-            final MetaData.Builder metaDataBuilder = MetaData.builder(recoveredState.metaData());
-            // automatically generate a UID for the metadata if we need to
-            metaDataBuilder.generateClusterUuidIfNeeded();
-
-            for (final IndexMetaData indexMetaData : recoveredState.metaData()) {
-                metaDataBuilder.put(indexMetaData, false);
-            }
-
-            return ClusterState.builder(currentState)
-                    .blocks(blocks)
-                    .metaData(metaDataBuilder)
-                    .build();
-        }
-
-    }
-
     class RecoverStateUpdateTask extends ClusterStateUpdateTask {
 
         @Override
         public ClusterState execute(final ClusterState currentState) {
             final ClusterState newState = Function.<ClusterState>identity()
-                    .andThen(state -> Updaters.closeBadIndices(state, indicesService))
-                    .andThen(state -> Updaters.upgradeAndArchiveUnknownOrInvalidSettings(state, clusterService.getClusterSettings()))
-                    .andThen(Updaters::recoverClusterBlocks)
-                    .andThen(Updaters::updateRoutingTable)
-                    .andThen(Updaters::removeStateNotRecoveredBlock)
+                    .andThen(ClusterStateUpdaters::updateRoutingTable)
+                    .andThen(ClusterStateUpdaters::removeStateNotRecoveredBlock)
                     .apply(currentState);
 
             return allocationService.reroute(newState, "state recovered");
@@ -392,8 +271,8 @@ public class GatewayService extends AbstractLifecycleComponent implements Cluste
             clusterService.submitStateUpdateTask("local-gateway-elected-state", new RecoverStateUpdateTask() {
                 @Override
                 public ClusterState execute(final ClusterState currentState) {
-                    final ClusterState updatedState = Updaters.mixCurrentStateAndRecoveredState(currentState, recoveredState);
-                    return super.execute(updatedState);
+                    final ClusterState updatedState = ClusterStateUpdaters.mixCurrentStateAndRecoveredState(currentState, recoveredState);
+                    return super.execute(ClusterStateUpdaters.recoverClusterBlocks(updatedState));
                 }
             });
         }
